pyproject.toml
[project]
name = "lms-chatcc"
version = "0.1.0"
description = "Streamlit web application for chatting with LM Studio"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit>=1.28.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0"
]
